moving attachSectionsToClasses WIP
